Move `task_list.py` into `dev` folder
